requirements.txt
# Local provider (HuggingFace / torch)
transformers>=4.43.0
torch>=2.1.0
accelerate>=0.30.0
sentencepiece>=0.2.0
safetensors>=0.4.3

# Gemini provider (Google hosted)
google-genai>=1.0.0

# Shared
pyyaml>=6.0.1
python-docx>=1.1.2
python-dotenv>=1.0.0
